refactor: 초기 필터링 데이터 세팅을 페이지가 담당하도록 변경
